fix(postgres,cockroachdb): include cockroachdb and move test to functional (#12238 review)
