+++ src/persistence/SubjectExecutor.ts
@@ -638,7 +638,9 @@ export class SubjectExecutor {
         }
 
         // Avoid concurrent queries on the same pg client; see #12238.
-        if (this.queryRunner.dataSource.driver.options.type === "postgres") {
+        // CockroachDB uses the pg package over a single connection too.
+        const driverType = this.queryRunner.dataSource.driver.options.type
+        if (driverType === "postgres" || driverType === "cockroachdb") {
             for (const subject of remainingSubjects) {
                 await updateSubject(subject)
             }
@@ -822,7 +824,9 @@ export class SubjectExecutor {
         }
 
         // Avoid concurrent queries on the same pg client; see #12238.
-        if (this.queryRunner.dataSource.options.type === "postgres") {
+        // CockroachDB uses the pg package over a single connection too.
+        const driverType = this.queryRunner.dataSource.options.type
+        if (driverType === "postgres" || driverType === "cockroachdb") {
             for (const subject of this.softRemoveSubjects) {
                 await softRemoveSubject(subject)
             }
@@ -931,7 +935,9 @@ export class SubjectExecutor {
         }
 
         // Avoid concurrent queries on the same pg client; see #12238.
-        if (this.queryRunner.dataSource.options.type === "postgres") {
+        // CockroachDB uses the pg package over a single connection too.
+        const driverType = this.queryRunner.dataSource.options.type
+        if (driverType === "postgres" || driverType === "cockroachdb") {
             for (const subject of this.recoverSubjects) {
                 await recoverSubject(subject)
             }
