+++ src/query-builder/relation-id/RelationIdLoader.ts
@@ -395,7 +395,9 @@ export class RelationIdLoader {
         }
 
         // Avoid concurrent queries on the same pg client; see #12238.
-        if (this.dataSource.options.type === "postgres") {
+        // CockroachDB uses the pg package over a single connection too.
+        const driverType = this.dataSource.options.type
+        if (driverType === "postgres" || driverType === "cockroachdb") {
             const results: RelationIdLoadResult[] = []
             for (const relationIdAttr of this.relationIdAttributes) {
                 results.push(await loadRelationIdAttribute(relationIdAttr))
