+++ test/functional/persistence/pg-concurrent-query-deprecation/entity/Tag.ts
@@ -1,0 +1,17 @@
+import { Entity } from "../../../../../src/decorator/entity/Entity"
+import { PrimaryGeneratedColumn } from "../../../../../src/decorator/columns/PrimaryGeneratedColumn"
+import { Column } from "../../../../../src/decorator/columns/Column"
+import { ManyToMany } from "../../../../../src/decorator/relations/ManyToMany"
+import { Post } from "./Post"
+
+@Entity()
+export class Tag {
+    @PrimaryGeneratedColumn()
+    id: number
+
+    @Column()
+    name: string
+
+    @ManyToMany(() => Post, (post) => post.tags)
+    posts: Post[]
+}
